==> 2_QualityChecks_resolved.Rmd <==
---
title: "Data Quality Checks"
author: "Carmen Rodriguez Cabrera and Emma Crenshaw"
date: "11/29/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(knitr)
library(tidyverse)
```


## Addressed 11/30/21 by Emma -  No HTML output

After webscrapping  we performed data checks to make sure the data obtained followed the right units (e.g., time variables), checked for typos, and validity of missingness. After these checks,   revisions were made to the webscrapping document.

```{r}
recipesdat<-read.csv(file = "recipes_final.csv")
str(recipesdat) # 1142 recipes
#Note: dataset is at the recipe level

```

```{r echo=TRUE, message=FALSE, warning=FALSE}
#Rating
summary(recipesdat$rating) #7 missing rating and so we remove missing
recipesdat %>% filter(is.na(rating)) %>% dplyr::select(recipe, servings)

#n_ratings, cook_time, prep_time, servings, n_ingredients, n_steps, cuisine

#Cook Time and Prep time
#How do we distinguish hour and minutes from cook or prep time?-- ASK EMMA 
recipesdat %>% filter(prep_time < 10) # we found typos on the website

recipesdat %>% filter(cook_time > 60) %>% dplyr::select(recipe, servings)
#Servings
summary(recipesdat$servings)
#16 missing this variable
#Max= 750--check, and minimum is 1

#recipes with very large serving sizes: 220 and 750
recipescheck<-recipesdat %>% filter(servings >60) %>% select(recipe)
#these are supposed to be 750 ml/ 3 cups and 220 ml/7 oz -- might need to drop


# Number of ingredients 
summary(recipesdat$n_ingredients)
table(recipesdat$n_ingredients)
#recipesdat %>% filter(n_ingredients ==84)

#Number of steps -- majority say 0
summary(recipesdat$n_steps)
table(recipesdat$n_steps)
nstepszero<-recipesdat %>% filter(n_steps==0)

#Taking a look at cuisine
recipesdat %>% group_by(cuisine) %>% summarise(n = n()) %>%ungroup() %>%  arrange(desc(n)) # find the 5 most common and then have an other category
```


# Summary of quality checks

-	Cook_time and prep_time:  we need to check values to make sure we have the same units. 

- Servings
    - Values 220 and 750 correspond to 750mL/3 cups and 220 mL/ 7 oz

- 	N_ingredients: recipe “CHICKEN CHOW MEIN,” says that it has 84 ingredients

-	N_steps:  878 recipes have 0 for the number of steps
