==> 5_Map Visualizations(final) - Aiza Malik.Rmd <==
---
title: "R Notebook"
output:
  html_document:
    df_print: paged
---
```{r}
load("/Users/aizamalik/Desktop/BST260/Project/BST260_project_data.RDa")
```

```{r}
library(dplyr)
library(tidyverse)
library(maps)
library(stringr)
```

Purpose: to visualize which cuisines, grouped by country, are the most interacted with (have the most ratings) and which ones are most common (number of recipes of a specific cuisine).

```{r}
head(data_final)
world_map = map_data("world")
```

First, I created a data frame 'commonentries' that counts the instances of an entry and arranges them in descending order by frequency. Then I created a dataframe 'commonratings' that counts the number of ratings each cuisine entry type has and arranges them in descending order. I combined these two data frames into a new dataframe 'commonvalues2'. 

```{r}
#commonstrings <- data_final %>% add_count(frequency = cuisine) %>% arrange(., desc(n), by_group = frequency)

#By number of entries in data frame:
commonentries <- data_final %>% group_by(cuisine) %>% summarise(count=n()) %>% arrange(desc(count))

#By number of ratings
commonratings <- data_final %>% group_by(cuisine) %>% filter(!is.na(n_ratings)) %>% summarise(count=sum(n_ratings)) %>% arrange(desc(count))

#Ratings normalized for number of recipes
commonvalues <- left_join(commonentries, commonratings, by = "cuisine") 
commonvalues2 <- commonvalues %>% group_by(cuisine) %>% mutate(normalized = sum(count.y/count.x)) %>% arrange(desc(normalized))
```

To simplify the mapping analysis I chose to focus on entries that gave us country-specific information. I excluded any regional entries and created a new data frame with the top 10 entries from 'commonvalues2'. I created a new column to have cuisine types match up with regions in 'world map' and then combined these data frames. 

```{r}
common <- commonvalues2 %>% filter(cuisine %in% c("Mexican", "Italian", "Chinese", "Thai", "French", "Indian", "Greek", "Vietnamese", "American", "Asian, Japanese")) %>% arrange(desc(count.x))
Countries = c("Mexico", "Italy", "China", "Thailand", "France", "India", "Greece", "Vietnam", "USA", "Japan")
common$region <- Countries
```
```{r}
world_map_data <- left_join(world_map, common, by = "region")
```

World maps mapping the cuisines with the most number of recipes and which cuisine was interacted with most frequently (measured by number of ratings for each cuisine)

The most common cuisine entries in order are: Mexican, Italian, Chinese, Thai, French, Indian, Greek, Vietnamese, American and Japanese

The most interacted-with cuisine entries in order are: Thai, Mexican, Chinese, Indian, Italian, Greek, Vietnamese, French, American, Japanese
```{r}
library(RColorBrewer)
map1 <- world_map_data %>% ggplot(aes(x = long, y = lat, group = group, fill = count.x)) +
   geom_polygon(color = "white") + scale_fill_gradientn(colors = brewer.pal(5, "Reds"),name="Number of Recipes") + 
   theme(panel.grid.major = element_blank(), 
         panel.background = element_blank(),
         axis.title = element_blank(), 
         axis.text = element_blank(),
         axis.ticks = element_blank()) + ggtitle("Most Popular Recipes")
map1
png("Map_Recipes.png", width = 700, height = 400)
map1
dev.off()
```
```{r}
library(RColorBrewer)
map2 <- world_map_data %>% ggplot(aes(x = long, y = lat, group = group, fill = count.y)) +
   geom_polygon(color = "white") + scale_fill_gradientn(colors = brewer.pal(5, "Reds"), name="Number of Ratings") +
   theme(panel.grid.major = element_blank(), 
         panel.background = element_blank(),
         axis.title = element_blank(), 
         axis.text = element_blank(),
         axis.ticks = element_blank()) + ggtitle("Most Interacted-With Recipes")
map2
png("Map_Ratings.png", width = 700, height = 400)
map2
dev.off()
```
